__DIR__ placeholder in config
